LayoutTests/http/tests/security/cross-frame-access-port-domain.html

<html>
<p>This test currently fails because we check the port and protocol even if document.domain is explicitly set (rdar://problem/5366437).</p> 
<iframe id="aFrame"></iframe>
<pre id="console"></pre>
<script>
    function log(s)
    {
        document.getElementById("console").appendChild(document.createTextNode(s + "\n"));
    }

    if (window.layoutTestController) {
        layoutTestController.dumpAsText();
        layoutTestController.waitUntilDone();
    }

    document.domain = "127.0.0.1";

    var targetWindow = frames[0];
    if (!targetWindow.document.body)
        log("FAIL: targetWindow started with no document, we won't know if the test passed or failed.");

    var iframe = document.getElementById('aFrame');
    iframe.src = "http://127.0.0.1:8080/security/resources/cross-frame-iframe-for-port-domain-test.html";

    var testDone = false;

    setTimeout(test, 1);

    setTimeout(function() {
        targetWindow.document.getElementById('accessMe').innerHTML = "PASS: Access allowed from frame with different port after explicitly setting document.domain.";
        if (!testDone) {
            log("Pass: Cross frame access to a different port was allowed.");
            if (window.layoutTestController)
                layoutTestController.notifyDone();
        }
    }, 5000);

    function test() {
        try {
            if (targetWindow.document.body) {
                setTimeout(test, 1);
                return;
            }
        } catch (e) {
        }

        log("Fail: Cross frame access to different port after after explicitly setting document.domain was denied!");
        testDone = true;
        if (window.layoutTestController)
            layoutTestController.notifyDone();
    }
</script>
</html>
